pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "groq==0.13.0",
    "instructor>=1.8.1",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.2",
    "uvicorn[standard]>=0.34.2",
]
